requirements.txt
pip
pytest==8.0.2
pytest-cov
requests
websocket-client
id-validator
names
bs4
jinja2
pytest-assume
pandas
openpyxl
qrcode[pil]
pyzbar
pyotp
pillow
selenium
PyYAML
allure-pytest
allure-python-commons
webdriver-manager
PyVirtualDisplay
Pillow
ImageHash
xvfbwrapper
opencv-python
pytest-testreport